refactor: modularize github actions fetcher
