+++ frontends/nextjs/src/components/misc/github/views/RunList.tsx
@@ -1,14 +1,11 @@
-import { Box, Stack, Typography } from '@mui/material'
+import { Stack } from '@mui/material'
 
-import { CheckCircle as SuccessIcon } from '@mui/icons-material'
+import { Card, CardContent, CardHeader } from '@/components/ui'
 
-import { Button, Card, CardContent, CardDescription, CardHeader, CardTitle, Skeleton } from '@/components/ui'
-
-import type { WorkflowRun } from '../types'
+import { Filters } from './run-list/Filters'
 import { RefreshControls } from './run-list/RefreshControls'
-import { RunItemCard } from './run-list/RunItemCard'
 import { RunListAlerts } from './run-list/RunListAlerts'
-import { RunListEmptyState } from './run-list/RunListEmptyState'
+import { RunTable } from './run-list/Table'
 import type { RunListProps } from './run-list/run-list.types'
 
 export function RunList({
@@ -39,32 +36,7 @@ export function RunList({
           alignItems={{ xs: 'flex-start', lg: 'center' }}
           justifyContent="space-between"
         >
-          <Stack spacing={1}>
-            <Typography variant="h4" fontWeight={700}>
-              GitHub Actions Monitor
-            </Typography>
-            <Typography color="text.secondary">
-              Repository:{' '}
-              <Box
-                component="code"
-                sx={{
-                  ml: 1,
-                  px: 1,
-                  py: 0.5,
-                  borderRadius: 1,
-                  bgcolor: 'action.hover',
-                  fontSize: '0.875rem',
-                }}
-              >
-                {repoLabel}
-              </Box>
-            </Typography>
-            {lastFetched && (
-              <Typography variant="caption" color="text.secondary">
-                Last fetched: {lastFetched.toLocaleString()}
-              </Typography>
-            )}
-          </Stack>
+          <Filters repoLabel={repoLabel} lastFetched={lastFetched} />
 
           <RefreshControls
             autoRefreshEnabled={autoRefreshEnabled}
@@ -86,57 +58,14 @@ export function RunList({
           summaryTone={summaryTone}
         />
 
-        <Card sx={{ borderWidth: 2, borderColor: 'divider' }}>
-          <CardHeader>
-            <Stack direction="row" spacing={1} alignItems="center" justifyContent="space-between">
-              <Stack direction="row" spacing={1} alignItems="center">
-                <SuccessIcon sx={{ color: 'success.main', fontSize: 24 }} />
-                <CardTitle>Recent Workflow Runs</CardTitle>
-              </Stack>
-              {isLoading && <Skeleton sx={{ width: 120, height: 12 }} />}
-            </Stack>
-            <CardDescription>Latest GitHub Actions runs with status and controls</CardDescription>
-          </CardHeader>
-
-          <CardContent>
-            {isLoading && !runs && (
-              <Stack spacing={2}>
-                <Skeleton sx={{ height: 96 }} />
-                <Skeleton sx={{ height: 96 }} />
-                <Skeleton sx={{ height: 96 }} />
-              </Stack>
-            )}
-
-            {runs && runs.length > 0 ? (
-              <Stack spacing={2}>
-                {runs.map((run: WorkflowRun) => (
-                  <RunItemCard
-                    key={run.id}
-                    run={run}
-                    getStatusColor={getStatusColor}
-                    onDownloadLogs={onDownloadLogs}
-                    isLoadingLogs={isLoadingLogs}
-                    selectedRunId={selectedRunId}
-                  />
-                ))}
-                <Box sx={{ textAlign: 'center', pt: 2 }}>
-                  <Button
-                    variant="outline"
-                    onClick={() => {
-                      if (!runs) return
-                      const jsonData = JSON.stringify(runs, null, 2)
-                      navigator.clipboard.writeText(jsonData)
-                    }}
-                  >
-                    Copy All as JSON
-                  </Button>
-                </Box>
-              </Stack>
-            ) : (
-              <RunListEmptyState isLoading={isLoading} />
-            )}
-          </CardContent>
-        </Card>
+        <RunTable
+          runs={runs}
+          isLoading={isLoading}
+          getStatusColor={getStatusColor}
+          onDownloadLogs={onDownloadLogs}
+          isLoadingLogs={isLoadingLogs}
+          selectedRunId={selectedRunId}
+        />
       </CardContent>
     </Card>
   )
