+++ frontends/nextjs/src/components/misc/github/workflows/hooks/useWorkflowRuns.ts
@@ -1,0 +1,12 @@
+import { useWorkflowRunsApi } from './useWorkflowRunsApi'
+import { useWorkflowRunsSelectors } from './useWorkflowRunsSelectors'
+
+export function useWorkflowRuns() {
+  const api = useWorkflowRunsApi()
+  const selectors = useWorkflowRunsSelectors(api.runs)
+
+  return {
+    ...api,
+    ...selectors,
+  }
+}
