+++ frontends/nextjs/src/components/misc/github/workflows/hooks/useWorkflowRunsApi.ts
@@ -1,11 +1,11 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { toast } from 'sonner'
 
-import { WorkflowRun, RepoInfo } from '../types'
+import { RepoInfo, WorkflowRun } from '../../types'
 
 const DEFAULT_REPO_LABEL = 'johndoe6345789/metabuilder'
 
-export function useWorkflowRuns() {
+export function useWorkflowRunsApi() {
   const [runs, setRuns] = useState<WorkflowRun[] | null>(null)
   const [isLoading, setIsLoading] = useState(false)
   const [error, setError] = useState<string | null>(null)
@@ -15,7 +15,10 @@ export function useWorkflowRuns() {
   const [secondsUntilRefresh, setSecondsUntilRefresh] = useState(30)
   const [autoRefreshEnabled, setAutoRefreshEnabled] = useState(true)
 
-  const repoLabel = repoInfo ? `${repoInfo.owner}/${repoInfo.repo}` : DEFAULT_REPO_LABEL
+  const repoLabel = useMemo(
+    () => (repoInfo ? `${repoInfo.owner}/${repoInfo.repo}` : DEFAULT_REPO_LABEL),
+    [repoInfo],
+  )
 
   const fetchRuns = useCallback(async () => {
     setIsLoading(true)
@@ -86,72 +89,6 @@ export function useWorkflowRuns() {
 
   const toggleAutoRefresh = () => setAutoRefreshEnabled((prev) => !prev)
 
-  const getStatusColor = (status: string, conclusion: string | null) => {
-    if (status === 'completed') {
-      if (conclusion === 'success') return 'success.main'
-      if (conclusion === 'failure') return 'error.main'
-      if (conclusion === 'cancelled') return 'text.secondary'
-    }
-    return 'warning.main'
-  }
-
-  const conclusion = useMemo(() => {
-    if (!runs || runs.length === 0) return null
-
-    const total = runs.length
-    const completed = runs.filter(r => r.status === 'completed').length
-    const successful = runs.filter(r => r.status === 'completed' && r.conclusion === 'success').length
-    const failed = runs.filter(r => r.status === 'completed' && r.conclusion === 'failure').length
-    const cancelled = runs.filter(r => r.status === 'completed' && r.conclusion === 'cancelled').length
-    const inProgress = runs.filter(r => r.status !== 'completed').length
-
-    const mostRecent = runs[0]
-    const mostRecentTimestamp = new Date(mostRecent.updated_at).getTime()
-    const timeThreshold = 5 * 60 * 1000
-    const recentWorkflows = runs.filter((run) => {
-      const runTimestamp = new Date(run.updated_at).getTime()
-      return mostRecentTimestamp - runTimestamp <= timeThreshold
-    })
-
-    const mostRecentPassed = recentWorkflows.every(
-      (run) => run.status === 'completed' && run.conclusion === 'success',
-    )
-    const mostRecentFailed = recentWorkflows.some(
-      (run) => run.status === 'completed' && run.conclusion === 'failure',
-    )
-    const mostRecentRunning = recentWorkflows.some((run) => run.status !== 'completed')
-
-    const successRate = total > 0 ? Math.round((successful / total) * 100) : 0
-    let health: 'healthy' | 'warning' | 'critical' = 'healthy'
-    if (failed / total > 0.3 || successRate < 60) {
-      health = 'critical'
-    } else if (failed > 0 || inProgress > 0) {
-      health = 'warning'
-    }
-
-    return {
-      total,
-      completed,
-      successful,
-      failed,
-      cancelled,
-      inProgress,
-      successRate,
-      health,
-      recentWorkflows,
-      mostRecentPassed,
-      mostRecentFailed,
-      mostRecentRunning,
-    }
-  }, [runs])
-
-  const summaryTone = useMemo(() => {
-    if (!conclusion) return 'warning'
-    if (conclusion.mostRecentPassed) return 'success'
-    if (conclusion.mostRecentFailed) return 'error'
-    return 'warning'
-  }, [conclusion])
-
   return {
     runs,
     isLoading,
@@ -164,8 +101,5 @@ export function useWorkflowRuns() {
     autoRefreshEnabled,
     toggleAutoRefresh,
     fetchRuns,
-    getStatusColor,
-    conclusion,
-    summaryTone,
   }
 }
